backend/pdfparser/requirements.txt
pandas
numpy
Pillow 
layoutparser
#"layoutparser[layoutmodels]"
#"layoutparser[ocr]"
pytesseract
df2image
opencv-python
